fix(codegen): export create-lib methods
